+++ site/src/pages/UsersPage/UsersPage.test.tsx
@@ -12,7 +12,7 @@ import {
 import { Language as usersXServiceLanguage } from "xServices/users/usersXService"
 import * as API from "../../api/api"
 import { Role } from "../../api/typesGenerated"
-import { Language as ResetPasswordDialogLanguage } from "../../components/Dialogs/ResetPasswordDialog/ResetPasswordDialog"
+import { Language as ResetPasswordDialogLanguage } from "./ResetPasswordDialog/ResetPasswordDialog"
 import { renderWithAuth } from "../../testHelpers/renderHelpers"
 import { server } from "../../testHelpers/server"
 import { Language as UsersPageLanguage, UsersPage } from "./UsersPage"
